Move ContextLogger to testing._utils
